+++ caribou/src/lib/heap.c
@@ -27,7 +27,7 @@
 #if !defined(__DOXYGEN__)
 __attribute__((weak))
 #endif
-void notify_heap_alloc(int16_t blocks)
+void notify_heap_alloc(int32_t blocks)
 {
 }
 
@@ -37,7 +37,7 @@ void notify_heap_alloc(int16_t blocks)
 #if !defined(__DOXYGEN__)
 __attribute__((weak))
 #endif
-void notify_heap_dealloc(int16_t blocks)
+void notify_heap_dealloc(int32_t blocks)
 {
 }
 
@@ -124,7 +124,7 @@ void heap_init(void* heap_base, void* heap_end)
 /**
 ** @brief the number of bytes per block
 */
-int16_t heap_block_size()
+int32_t heap_block_size()
 {
 	return HEAP_BLOCK_SIZE();
 }
@@ -132,7 +132,7 @@ int16_t heap_block_size()
 /**
 ** @brief total up all the blocks used.
 */
-int16_t heap_blocks_allocated()
+int32_t heap_blocks_allocated()
 {
 	return HEAP_BLOCKS_ALLOCATED();
 }
